requirements.txt
appdirs
APScheduler
arrow
beautifulsoup4
bleach
certifi
cffi
chardet
cheroot
CherryPy
configobj
cryptography
distro
dnspython
gevent
gntp
greenlet
httpagentparser
idna
ipwhois
IPy
jaraco.classes
jaraco.collections
jaraco.functools
jaraco.text
jellyfin-apiclient-python
logutils
Mako
MarkupSafe
more-itertools
musicbrainzngs
packaging
paho-mqtt
portend
profilehooks
pycparser
PyJWT
pyOpenSSL
pyparsing
python-dateutil
pytz
requests
six
soupsieve
tempora
twitter
tzlocal
urllib3
webencodings
websocket-client
xmltodict
zc.lockfile
zope.event
zope.interface
cloudinary